# api-service/api/docs/swagger.yaml
definitions:
  booking.AvailabilityR:
    properties:
      end_time:
        type: string
      start_time:
        type: string
    type: object
...
  booking.CreateProvidersRequest:
    properties:
      availabilitys:
        items:
          $ref: '#/definitions/booking.AvailabilityR'
        type: array
      company_name:
        type: string
      location:
        $ref: '#/definitions/booking.Location'
      service_id:
        items:
          $ref: '#/definitions/booking.ServiceId'
        type: array
      user_id:
        type: string
    type: object
  booking.CreateProvidersResponse:
    properties:
      _id:
        type: string
      availabilitys:
        items:
          $ref: '#/definitions/booking.AvailabilityR'
        type: array
      company_name:
        type: string
      crerated_at:
        type: string
      location:
        $ref: '#/definitions/booking.Location'
      service_id:
        items:
          $ref: '#/definitions/booking.ServiceId'
        type: array
      user_id:
        type: string
    type: object
  booking.CreateReviewRequest:
    properties:
      booking_id:
        type: string
      comment:
        type: string
      procvider_id:
        type: string
      rating:
        type: number
      user_id:
        type: string
    type: object
  booking.CreateReviewResponse:
    properties:
      booking_id:
        type: string
      comment:
        type: string
      created_at:
        type: string
      procvider_id:
        type: string
      rating:
        type: number
      user_id:
        type: string
    type: object
...
  booking.DeleteProvidersResponse:
    properties:
      message:
        type: string
    type: object
  booking.DeleteReviewResponse:
    properties:
      message:
        type: string
    type: object
...
  booking.GetProvidersResponse:
    properties:
      _id:
        type: string
      availabilitys:
        items:
          $ref: '#/definitions/booking.AvailabilityR'
        type: array
      company_name:
        type: string
      crerated_at:
        type: string
      location:
        $ref: '#/definitions/booking.Location'
      service_id:
        items:
          $ref: '#/definitions/booking.ServiceId'
        type: array
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  booking.ListProviderR:
    properties:
      _id:
        type: string
      availabilitys:
        items:
          $ref: '#/definitions/booking.AvailabilityR'
        type: array
      company_name:
        type: string
      crerated_at:
        type: string
      location:
        $ref: '#/definitions/booking.Location'
      service_id:
        items:
          $ref: '#/definitions/booking.ServiceId'
        type: array
      user_id:
        type: string
    type: object
  booking.ListProvidersResponse:
    properties:
      listpriders:
        items:
          $ref: '#/definitions/booking.ListProviderR'
        type: array
    type: object
  booking.ListReviewR:
    properties:
      _id:
        type: string
      booking_id:
        type: string
      comment:
        type: string
      created_at:
        type: string
      procvider_id:
        type: string
      rating:
        type: number
      user_id:
        type: string
    type: object
  booking.ListReviewsResponse:
    properties:
      listreviews:
        items:
          $ref: '#/definitions/booking.ListReviewR'
        type: array
    type: object
...
  booking.Location:
    properties:
      city:
        type: string
      country:
        type: string
    type: object
...
  booking.SearchProviderR:
    properties:
      _id:
        type: string
      availabilitys:
        items:
          $ref: '#/definitions/booking.AvailabilityR'
        type: array
      company_name:
        type: string
      crerated_at:
        type: string
      location:
        $ref: '#/definitions/booking.Location'
      service_id:
        items:
          $ref: '#/definitions/booking.ServiceId'
        type: array
      user_id:
        type: string
    type: object
  booking.SearchProvidersResponse:
    properties:
      provider:
        items:
          $ref: '#/definitions/booking.SearchProviderR'
        type: array
    type: object
...
  booking.ServiceId:
    properties:
      _id:
        type: string
    type: object
  booking.UpdateProvidersRequest:
    properties:
      _id:
        type: string
      availabilitys:
        items:
          $ref: '#/definitions/booking.AvailabilityR'
        type: array
      company_name:
        type: string
      location:
        $ref: '#/definitions/booking.Location'
      user_id:
        type: string
    type: object
  booking.UpdateProvidersResponse:
    properties:
      _id:
        type: string
      availabilitys:
        items:
          $ref: '#/definitions/booking.AvailabilityR'
        type: array
      company_name:
        type: string
      crerated_at:
        type: string
      location:
        $ref: '#/definitions/booking.Location'
      service_id:
        items:
          $ref: '#/definitions/booking.ServiceId'
        type: array
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  booking.UpdateReviewRequest:
    properties:
      _id:
        type: string
      comment:
        type: string
      procvider_id:
        type: string
      rating:
        type: number
      user_id:
        type: string
    type: object
  booking.UpdateReviewResponse:
    properties:
      booking_id:
        type: string
      comment:
        type: string
      procvider_id:
        type: string
      rating:
        type: number
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
info:
  contact: {}
  title: On-Demand API
  version: "1.0"
paths:
...
  /provider/providers:
    get:
      consumes:
      - application/json
      description: List Providers
      parameters:
      - description: limit
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ListProvidersResponse'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: List Providers
      tags:
      - Providers
    post:
      consumes:
      - application/json
      description: Create Providers
      parameters:
      - description: providers
        in: body
        name: providers
        required: true
        schema:
          $ref: '#/definitions/booking.CreateProvidersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.CreateProvidersResponse'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Create Providers
      tags:
      - Providers
  /provider/providers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Providers
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.DeleteProvidersResponse'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Delete Providers
      tags:
      - Providers
    get:
      consumes:
      - application/json
      description: Get Providers
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.GetProvidersResponse'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get Providers
      tags:
      - Providers
    put:
      consumes:
      - application/json
      description: Update Providers
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: providers
        in: body
        name: providers
        schema:
          $ref: '#/definitions/booking.UpdateProvidersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.UpdateProvidersResponse'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Update Providers
      tags:
      - Providers
  /provider/providers/search:
    get:
      consumes:
      - application/json
      description: Search Providers
      parameters:
      - description: user_id
        in: query
        name: user_id
        type: string
      - description: company_name
        in: query
        name: company_name
        type: string
      - description: location
        in: query
        name: location
        type: string
      - collectionFormat: csv
        description: availability_as
        in: query
        items:
          type: string
        name: availability_as
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.SearchProvidersResponse'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Search Providers
      tags:
      - Providers
  /review/reviews:
    get:
      consumes:
      - application/json
      description: List Reviews
      parameters:
      - description: limit
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ListReviewsResponse'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: List Reviews
      tags:
      - Review
    post:
      consumes:
      - application/json
      description: Create Review
      parameters:
      - description: review
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/booking.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.CreateReviewResponse'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Create Review
      tags:
      - Review
  /review/reviews/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Review
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.DeleteReviewResponse'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Delete Review
      tags:
      - Review
    put:
      consumes:
      - application/json
      description: Update Review
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: review
        in: body
        name: review
        schema:
          $ref: '#/definitions/booking.UpdateReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.UpdateReviewResponse'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Update Review
      tags:
      - Review
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"